requirements.txt
torch>=2.0.0
numpy>=1.21.0
gymnasium>=0.29.0
pandas>=1.5.0
matplotlib>=3.5.0
tensorboard>=2.12.0
tqdm>=4.65.0
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
pyyaml>=6.0.0
seaborn>=0.12.0
